requirements.txt
aniso8601==9.0.1
blinker==1.9.0
click==8.1.7
dnspython==2.7.0
Flask==3.1.0
Flask-RESTful==0.3.10
gunicorn==23.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==3.0.2
packaging==24.2
pymongo==4.6.0
python-dotenv==1.0.1
pytz==2024.2
six==1.16.0
Werkzeug==3.1.3
